graph_coarsening/graph_classification/requirements.txt
alabaster==0.7.12
allennlp==0.8.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==20.2.0
aws-sam-translator==1.28.1
aws-xray-sdk==2.6.0
awscli==1.18.169
Babel==2.8.0
backcall==0.2.0
backports.weakref==1.0rc1
bleach==1.5.0
blessings==1.6
blis==0.2.4
boto==2.49.0
boto3==1.14.48
botocore==1.17.48
cached-property==1.5.2
catalogue==1.0.0
certifi==2021.10.8
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046052501/work
cfn-lint==0.39.0
chardet==3.0.4
click @ file:///home/conda/feedstock_root/build_artifacts/click_1635822578576/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conllu==0.11
cryptography==3.2.1
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
cymem==2.0.3
Cython==0.29.21
debugpy==1.5.1
decorator==4.4.2
deepwalk==1.0.3
defusedxml==0.7.1
docker==4.3.1
docopt==0.6.2
docutils==0.15.2
easydict==1.9
ecdsa==0.14.1
edit-distance==1.0.4
editdistance==0.5.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
entrypoints==0.3
filelock==3.0.12
flaky==3.7.0
Flask==1.1.2
Flask-Cors==3.0.9
ftfy==5.8
future @ file:///home/conda/feedstock_root/build_artifacts/future_1635819510385/work
gensim==3.5.0
gevent==20.9.0
graph-coarsening @ file:///home/liangy7/Research/graph_coarsening/graph-coarsening
graphvite==0.2.2
greenlet==0.4.17
h5py==3.0.0
html5lib==0.9999999
idna==2.10
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1594044661732/work
imagesize==1.2.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1636431542280/work
inflect==4.1.0
iniconfig==1.1.1
ipykernel==6.7.0
ipython==7.31.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.2
jmespath==0.10.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
jsondiff==1.2.0
jsonlines==1.2.0
jsonnet==0.16.0
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1635836699543/work
MarkupSafe==1.1.1
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1636786662993/work
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.6.0
moto==1.3.16
msgpack==0.5.6
murmurhash==1.0.2
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
NetLSD==1.0.2
networkx==2.5
neuralcoref==4.0
nltk @ file:///home/conda/feedstock_root/build_artifacts/nltk_1633955089856/work
notebook==6.4.8
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1636145336524/work
numpydoc==1.1.0
olefile==0.46
overrides==3.1.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1636558793805/work
pipreqs==0.4.10
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
progressbar==2.5
prometheus-client==0.13.0
prompt-toolkit==3.0.24
protobuf==3.13.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1635822683042/work
ptyprocess==0.7.0
py==1.9.0
pyasn1==0.4.8
pybind11 @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1635827735553/work
pybind11-global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1635827735553/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyemd==0.5.1
Pygments==2.7.2
PyGSP==0.5.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1636757021002/work
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
pytest==6.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-jose==3.2.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
pytz==2020.1
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1636139801027/work
pyzmq==22.3.0
rapidfuzz==0.3.0
regex @ file:///home/conda/feedstock_root/build_artifacts/regex_1636504130952/work
requests==2.24.0
responses==0.12.0
rsa==4.5
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.22.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1604304773994/work
Send2Trash==1.8.0
sentencepiece==0.1.91
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==2.1.0
snowballstemmer==2.0.0
sortedcontainers==2.4.0
spacy==2.1.9
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
srsly==1.0.2
sshpubkeys==3.1.0
tensorboardX==2.1
terminado==0.12.1
testpath==0.5.0
thinc==7.0.8
threadpoolctl==2.1.0
tokenizers==0.9.2
toml==0.10.2
torch @ file:///home/conda/feedstock_root/build_artifacts/pytorch-recipe_1610938746276/work
torchtext==0.7.0
torchvision @ file:///home/conda/feedstock_root/build_artifacts/torchvision-split_1615654770770/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1635819587639/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1632160078689/work
traitlets==5.1.1
transformers==3.4.0
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1632313171031/work
Unidecode==1.1.1
urllib3==1.25.10
wasabi==0.8.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
word2number==1.1
wrapt==1.12.1
xmltodict==0.12.0
yarg==0.1.9
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1633302054558/work
zope.event==4.5.0
zope.interface==5.1.2
